The submit button acutally posts now!!!
SO the wizard now has the aility to post to the server and sucessfully
does so.

I couldn't manage to get the APICalls/RestClient to work with the slim
framework
I ended up adding my own post code (POST.java) to work with the
framework in the way we were expecting.
We can talk more about this at our meeting this weekend.

Things TODO for the wizard:
Enable location finding.
Enable marker on map.
Add picker for time selection.
Fix toast issue.
Enabel threading to display a thinking symbol while post is happening.